edits

+++ src/Provider.php
@@ -2,6 +2,8 @@
 
 namespace FII;
 
+use FII\Models\Course;
+use FII\Models\CourseReview;
 use Requests_Session;
 
 class Provider
@@ -16,14 +18,25 @@ class Provider
         $this->cache = $cache;
     }
 
-    public function persist($courses)
+
+    public function persist($data)
     {
-        //clean the records
-        Course::truncate();
-        $courses->each(function($course){
-            $course->save();
-        });
-        return $this;
+        $class = get_class($data);
+        if(preg_match('/course/i', $class)){
+            Course::truncate();
+            $data->each(function($course){
+                $course->save();
+            });
+            return $this;
+        }
+
+        if(preg_match('/review/i', $class)) {
+            CourseReview::truncate();
+            $data->each(function($review){
+                $review->save();
+            });
+            return $this;
+        }
     }
 
 
@@ -62,8 +75,10 @@ class Provider
     }
 
 
-    public function cache($data, $key)
+    public function cache($data)
     {
+        $key = forward_static_call_array([$data, 'cacheKey'], []);
         $this->cache->put($key, $data, CACHE_WEEK);
+        return $this;
     }
 }
